refactor: reorganiza DTOs em packeges
