+++ ArchSpot-BackEnd/src/main/java/com/archspot/ArchSpot_BackEnd/controllers/AuthController.java
@@ -1,8 +1,8 @@
 package com.archspot.ArchSpot_BackEnd.controllers;
 
-import com.archspot.ArchSpot_BackEnd.dtos.LoginRequestDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.UserDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.UserCreateDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.auth.LoginRequestDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.user.UserCreateDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.user.UserDTO;
 import com.archspot.ArchSpot_BackEnd.entities.User;
 import com.archspot.ArchSpot_BackEnd.security.AuthUser;
 import com.archspot.ArchSpot_BackEnd.security.TokenService;
