+++ ArchSpot-BackEnd/src/main/java/com/archspot/ArchSpot_BackEnd/controllers/DirectoryController.java
@@ -1,8 +1,8 @@
 package com.archspot.ArchSpot_BackEnd.controllers;
 
-import com.archspot.ArchSpot_BackEnd.dtos.DirectoryCreateDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.DirectoryDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.DocumentDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.diretory.DirectoryCreateDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.diretory.DirectoryDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.document.DocumentDTO;
 import com.archspot.ArchSpot_BackEnd.services.DirectoryService;
 import com.archspot.ArchSpot_BackEnd.services.DocumentService;
