+++ ArchSpot-BackEnd/src/main/java/com/archspot/ArchSpot_BackEnd/controllers/InstallmentController.java
@@ -1,7 +1,7 @@
 package com.archspot.ArchSpot_BackEnd.controllers;
 
-import com.archspot.ArchSpot_BackEnd.dtos.InstallmentRequestDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.InstallmentResponseDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.installment.InstallmentRequestDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.installment.InstallmentResponseDTO;
 import com.archspot.ArchSpot_BackEnd.enums.PaymentMethod;
 import com.archspot.ArchSpot_BackEnd.enums.PaymentStatus;
 import com.archspot.ArchSpot_BackEnd.services.InstallmentService;
