+++ ArchSpot-BackEnd/src/main/java/com/archspot/ArchSpot_BackEnd/controllers/ProjectController.java
@@ -20,16 +20,16 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
-import com.archspot.ArchSpot_BackEnd.dtos.AlbumCreateDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.AlbumDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.DirectoryCreateDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.DirectoryDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.InstallmentResponseDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.PhaseDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.ProjectRequestDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.ProjectResponseDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.UserProjectRequestDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.UserProjectResponseDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.album.AlbumCreateDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.album.AlbumDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.diretory.DirectoryCreateDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.diretory.DirectoryDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.installment.InstallmentResponseDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.phase.PhaseDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.project.ProjectRequestDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.project.ProjectResponseDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.userproject.UserProjectRequestDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.userproject.UserProjectResponseDTO;
 import com.archspot.ArchSpot_BackEnd.enums.DirectoryType;
 import com.archspot.ArchSpot_BackEnd.enums.PaymentStatus;
 import com.archspot.ArchSpot_BackEnd.enums.UserRole;
