+++ ArchSpot-BackEnd/src/main/java/com/archspot/ArchSpot_BackEnd/dtos/comment/CommentCreateDTO.java
@@ -1,4 +1,4 @@
-package com.archspot.ArchSpot_BackEnd.dtos;
+package com.archspot.ArchSpot_BackEnd.dtos.comment;
 
 import lombok.*;
