+++ ArchSpot-BackEnd/src/main/java/com/archspot/ArchSpot_BackEnd/dtos/diretory/DirectoryCreateDTO.java
@@ -1,4 +1,4 @@
-package com.archspot.ArchSpot_BackEnd.dtos;
+package com.archspot.ArchSpot_BackEnd.dtos.diretory;
 
 import com.archspot.ArchSpot_BackEnd.enums.DirectoryType;
 import lombok.*;
