+++ ArchSpot-BackEnd/src/main/java/com/archspot/ArchSpot_BackEnd/dtos/installment/InstallmentRequestDTO.java
@@ -1,4 +1,4 @@
-package com.archspot.ArchSpot_BackEnd.dtos;
+package com.archspot.ArchSpot_BackEnd.dtos.installment;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
@@ -6,12 +6,12 @@ import java.time.LocalDate;
 import com.archspot.ArchSpot_BackEnd.enums.PaymentMethod;
 import com.archspot.ArchSpot_BackEnd.enums.PaymentStatus;
 
-public record InstallmentResponseDTO(
-    Long id,
+public record InstallmentRequestDTO(
     LocalDate estimatedPaymentDate,
     LocalDate realPaymentDate,
     PaymentMethod paymentMethod,
     PaymentStatus paymentStatus,
     BigDecimal amount,
     String description,
-    Long projectId) {}
+    Long projectId) {
+}
