+++ ArchSpot-BackEnd/src/main/java/com/archspot/ArchSpot_BackEnd/dtos/photo/PhotoUploadDTO.java
@@ -1,4 +1,4 @@
-package com.archspot.ArchSpot_BackEnd.dtos;
+package com.archspot.ArchSpot_BackEnd.dtos.photo;
 
 import lombok.*;
