+++ ArchSpot-BackEnd/src/main/java/com/archspot/ArchSpot_BackEnd/services/PhaseService.java
@@ -11,9 +11,9 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 import com.archspot.ArchSpot_BackEnd.enums.*;
-import com.archspot.ArchSpot_BackEnd.dtos.PhaseCreateDTO;
-import com.archspot.ArchSpot_BackEnd.dtos.PhaseDTO;
 import com.archspot.ArchSpot_BackEnd.entities.Project;
+import com.archspot.ArchSpot_BackEnd.dtos.phase.PhaseCreateDTO;
+import com.archspot.ArchSpot_BackEnd.dtos.phase.PhaseDTO;
 import com.archspot.ArchSpot_BackEnd.entities.Phase;
 import com.archspot.ArchSpot_BackEnd.repositories.PhaseRepository;
 import com.archspot.ArchSpot_BackEnd.repositories.ProjectRepository;
